refactor(index): split files
